backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "fastapi[standard]>=0.135.3",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.11",
    "pwdlib[argon2]>=0.3.0",
    "pytest>=9.0.3",
    "python-jose[cryptography]>=3.5.0",
    "sqlmodel>=0.0.38",
]


